# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MEET Automotive Community API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  viewer: "viewer",
  participant: "participant",
  organizer: "organizer",
} as const;

export type UserViewerSilhouette =
  | (typeof UserViewerSilhouette)[keyof typeof UserViewerSilhouette]
  | null;

export const UserViewerSilhouette = {
  bicycle: "bicycle",
  scooter: "scooter",
  skateboard: "skateboard",
  cart: "cart",
} as const;

export interface User {
  id: number;
  telegramId: string;
  username: string;
  displayName: string;
  avatarUrl?: string | null;
  role: UserRole;
  organizationName?: string | null;
  contactLink?: string | null;
  viewerSilhouette?: UserViewerSilhouette;
  interestCategories: string[];
  onboardingComplete: boolean;
  createdAt: string;
}

export type OnboardingRequestRole =
  (typeof OnboardingRequestRole)[keyof typeof OnboardingRequestRole];

export const OnboardingRequestRole = {
  viewer: "viewer",
  participant: "participant",
  organizer: "organizer",
} as const;

export type OnboardingRequestViewerSilhouette =
  | (typeof OnboardingRequestViewerSilhouette)[keyof typeof OnboardingRequestViewerSilhouette]
  | null;

export const OnboardingRequestViewerSilhouette = {
  bicycle: "bicycle",
  scooter: "scooter",
  skateboard: "skateboard",
  cart: "cart",
} as const;

export interface OnboardingRequest {
  telegramId: string;
  username: string;
  displayName: string;
  avatarUrl?: string | null;
  role: OnboardingRequestRole;
  viewerSilhouette?: OnboardingRequestViewerSilhouette;
  organizationName?: string | null;
  contactLink?: string | null;
  interestCategories?: string[];
}

export type EventCategory = (typeof EventCategory)[keyof typeof EventCategory];

export const EventCategory = {
  motorsport: "motorsport",
  exhibition: "exhibition",
  cruise: "cruise",
  club: "club",
} as const;

export type EventStatus = (typeof EventStatus)[keyof typeof EventStatus];

export const EventStatus = {
  upcoming: "upcoming",
  ongoing: "ongoing",
  finished: "finished",
  cancelled: "cancelled",
} as const;

export interface Event {
  id: number;
  title: string;
  description?: string | null;
  category: EventCategory;
  subcategories: string[];
  organizerId: number;
  organizerName: string;
  organizerContact?: string | null;
  date: string;
  endDate?: string | null;
  location: string;
  lat?: number | null;
  lng?: number | null;
  maxParticipants?: number | null;
  isPrivate: boolean;
  coverImageUrl?: string | null;
  status: EventStatus;
  applicationsCount: number;
  viewersCount: number;
  createdAt: string;
}

export type CreateEventRequestCategory =
  (typeof CreateEventRequestCategory)[keyof typeof CreateEventRequestCategory];

export const CreateEventRequestCategory = {
  motorsport: "motorsport",
  exhibition: "exhibition",
  cruise: "cruise",
  club: "club",
} as const;

export interface CreateEventRequest {
  title: string;
  description?: string | null;
  category: CreateEventRequestCategory;
  subcategories: string[];
  date: string;
  endDate?: string | null;
  location: string;
  lat?: number | null;
  lng?: number | null;
  maxParticipants?: number | null;
  isPrivate: boolean;
  coverImageUrl?: string | null;
}

export type ApplicationType =
  (typeof ApplicationType)[keyof typeof ApplicationType];

export const ApplicationType = {
  participant: "participant",
  viewer: "viewer",
} as const;

export type ApplicationStatus =
  (typeof ApplicationStatus)[keyof typeof ApplicationStatus];

export const ApplicationStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Application {
  id: number;
  eventId: number;
  eventTitle?: string | null;
  userId: number;
  userName?: string | null;
  userAvatarUrl?: string | null;
  carId?: number | null;
  carMake?: string | null;
  carModel?: string | null;
  type: ApplicationType;
  status: ApplicationStatus;
  comment?: string | null;
  createdAt: string;
}

export type ApplyRequestType =
  (typeof ApplyRequestType)[keyof typeof ApplyRequestType];

export const ApplyRequestType = {
  participant: "participant",
  viewer: "viewer",
} as const;

export interface ApplyRequest {
  carId?: number | null;
  type: ApplyRequestType;
  comment?: string | null;
}

export type UpdateApplicationRequestStatus =
  (typeof UpdateApplicationRequestStatus)[keyof typeof UpdateApplicationRequestStatus];

export const UpdateApplicationRequestStatus = {
  approved: "approved",
  rejected: "rejected",
} as const;

export interface UpdateApplicationRequest {
  status: UpdateApplicationRequestStatus;
}

export interface Car {
  id: number;
  userId: number;
  make: string;
  model: string;
  year?: number | null;
  color?: string | null;
  photoUrl?: string | null;
  aiStyledImageUrl?: string | null;
  categories: string[];
  isPrimary: boolean;
  createdAt: string;
}

export interface AddCarRequest {
  make: string;
  model: string;
  year?: number | null;
  color?: string | null;
  photoUrl?: string | null;
  categories?: string[];
  isPrimary?: boolean;
}

export interface Club {
  id: number;
  name: string;
  description?: string | null;
  category: string;
  subcategory?: string | null;
  logoUrl?: string | null;
  contactLink?: string | null;
  membersCount: number;
  createdAt: string;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  event_reminder: "event_reminder",
  application_approved: "application_approved",
  application_rejected: "application_rejected",
  new_event: "new_event",
  event_cancelled: "event_cancelled",
} as const;

export interface Notification {
  id: number;
  userId: number;
  type: NotificationType;
  title: string;
  message: string;
  eventId?: number | null;
  isRead: boolean;
  createdAt: string;
}

export type GetEventsParams = {
  category?: string;
  lat?: number;
  lng?: number;
  radius?: number;
  dateFrom?: string;
  dateTo?: string;
  search?: string;
};

export type GetClubsParams = {
  make?: string;
  category?: string;
};