# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MEET Automotive Community API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/{eventId}/applications:
    get:
      operationId: getEventApplications
      tags: [applications]
      summary: Get applications for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: applyToEvent
      tags: [applications]
      summary: Apply to participate in event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyRequest"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /events/{eventId}/applications/{applicationId}:
    put:
      operationId: updateApplication
      tags: [applications]
      summary: Update application status (organizer)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
        - name: applicationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationRequest"
      responses:
        "200":
          description: Application updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
    delete:
      operationId: cancelApplication
      tags: [applications]
      summary: Cancel/delete application
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
        - name: applicationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application cancelled
  /cars:
    get:
      operationId: getMyCars
      tags: [cars]
      summary: Get my garage (cars)
      responses:
        "200":
          description: List of cars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Car"
    post:
      operationId: addCar
      tags: [cars]
      summary: Add a car to garage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCarRequest"
      responses:
        "201":
          description: Car added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
  /cars/{carId}:
    put:
      operationId: updateCar
      tags: [cars]
      summary: Update car details
      parameters:
        - name: carId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCarRequest"
      responses:
        "200":
          description: Car updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
    delete:
      operationId: deleteCar
      tags: [cars]
      summary: Remove car from garage
      parameters:
        - name: carId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Car removed
...
components:
  schemas:
...
    Application:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        eventTitle:
          type: string
          nullable: true
        userId:
          type: integer
        userName:
          type: string
          nullable: true
        userAvatarUrl:
          type: string
          nullable: true
        carId:
          type: integer
          nullable: true
        carMake:
          type: string
          nullable: true
        carModel:
          type: string
          nullable: true
        type:
          type: string
          enum: [participant, viewer]
        status:
          type: string
          enum: [pending, approved, rejected]
        comment:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - eventId
        - userId
        - type
        - status
        - createdAt
    ApplyRequest:
      type: object
      properties:
        carId:
          type: integer
          nullable: true
        type:
          type: string
          enum: [participant, viewer]
        comment:
          type: string
          nullable: true
      required:
        - type
    UpdateApplicationRequest:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected]
      required:
        - status
    Car:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        make:
          type: string
        model:
          type: string
        year:
          type: integer
          nullable: true
        color:
          type: string
          nullable: true
        photoUrl:
          type: string
          nullable: true
        aiStyledImageUrl:
          type: string
          nullable: true
        categories:
          type: array
          items:
            type: string
        isPrimary:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - userId
        - make
        - model
        - isPrimary
        - categories
        - createdAt
    AddCarRequest:
      type: object
      properties:
        make:
          type: string
        model:
          type: string
        year:
          type: integer
          nullable: true
        color:
          type: string
          nullable: true
        photoUrl:
          type: string
          nullable: true
        categories:
          type: array
          items:
            type: string
        isPrimary:
          type: boolean
      required:
        - make
        - model
...