# lib/api-zod/src/generated/types/applyRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MEET Automotive Community API
 * OpenAPI spec version: 0.1.0
 */
import type { ApplyRequestType } from "./applyRequestType";

export interface ApplyRequest {
  carId?: number | null;
  type: ApplyRequestType;
  comment?: string | null;
}